addMitarbeiter

+++ Backend/src/main/java/com/nsj/gfos/award/resources/ArbeitsgruppenResource.java
@@ -124,16 +124,27 @@ public class ArbeitsgruppenResource {
 		String leiter = attributes[2].split("=")[1];
 		if (bezeichnung == "")
 			return JsonHandler.fehler("Die Bezeichnung der Arbeitsgruppe ist leer.");
+		if (Utils.checkIfArbeitsgruppeExistsFromBezeichnung(bezeichnung))
+			return JsonHandler.fehler("Die Bezeichnung der Arbeitsgruppe ist bereits vergeben.");
 		if(!Utils.checkIfMitarbeiterExists(leiter))
 			return JsonHandler.fehler("Der Mitarbeiter, der Leiter werden soll, existiert nicht.");
 		if (!SessionHandler.checkSessionID(auth))
 			return JsonHandler.fehler("SessionID ist ungültig.");
 		// TODO checkRights + die des Leiters
-		String sqlStmt = "INSERT INTO gfos.arbeitsgruppe (ArbeitsgruppenID, Bezeichnung, Leiter) Values ('"+ Utils.createArbeitsgruppenID() +"', '" + bezeichnung + "', '" + leiter + "');";
+		String arbeitsgruppenID = Utils.createArbeitsgruppenID();
+		String sqlStmt = "INSERT INTO gfos.arbeitsgruppe (ArbeitsgruppenID, Bezeichnung, Leiter) Values ('"+ arbeitsgruppenID +"', '" + bezeichnung + "', '" + leiter + "');";
 		try {
 			int rs = QueryHandler.update(sqlStmt);
 			if(rs == 0)
-				return JsonHandler.fehler("Das Einfügen konnte nicht durchgeführt werden.");
+				return JsonHandler.fehler("Das Einfügen der Arbeitsgruppe konnte nicht durchgeführt werden.");
+		} catch (SQLException e) {
+			return JsonHandler.fehler(e.toString());
+		}
+		sqlStmt = "INSERT INTO gfos.arbeitsgruppenteilnahme (ArbeitsgruppenID, Mitarbeiter) Values ('"+ arbeitsgruppenID +"', '" + leiter + "');";
+		try {
+			int rs = QueryHandler.update(sqlStmt);
+			if(rs == 0)
+				return JsonHandler.fehler("Das Einfügen des Leiters konnte nicht durchgeführt werden.");
 			return JsonHandler.erfolg("Arbeitsgruppe wurde erfolgreich erstellt.");
 		} catch (SQLException e) {
 			return JsonHandler.fehler(e.toString());
@@ -162,6 +173,8 @@ public class ArbeitsgruppenResource {
 			return JsonHandler.fehler("ID ist ungültig");
 		if (!SessionHandler.checkSessionID(auth))
 			return JsonHandler.fehler("SessionID ist ungültig.");
+		if(!Utils.checkIfArbeitsgruppeExistsFromID(id))
+			return JsonHandler.fehler("Arbeitsgruppe existiert nicht.");
 		// TODO checkRights
 		String sqlStmt = "DELETE FROM gfos.arbeitsgruppenteilnahme WHERE gfos.arbeitsgruppenteilnahme.ArbeitsgruppenID = \"" + id + "\";";
 		try {
@@ -209,6 +222,8 @@ public class ArbeitsgruppenResource {
 			return JsonHandler.fehler("Ungültige Personalnummer.");
 		if(id.length() != 12)
 			return JsonHandler.fehler("Ungültige ArbeitsgruppenID.");
+		if(!Utils.checkIfMitarbeitserIsInArbeitsgruppe(id, pn))
+			return JsonHandler.fehler("Der Mitarbeiter befindet sich nicht in der Arbeitsgruppe.");
 		// TODO checkRights
 		if (!SessionHandler.checkSessionID(auth))
 			return JsonHandler.fehler("SessionID ist ungültig.");
@@ -234,6 +249,41 @@ public class ArbeitsgruppenResource {
 		}
 	}
 	
-	
+	@GET
+	@Path("addMitarbeiter{attributes}")
+	@Produces(MediaType.APPLICATION_JSON)
+	public static String addMitarbeiter(@PathParam("attributes") String query) {
+		query = query.substring(1);
+		String[] attributes = query.split("&");
+		if (attributes.length != 3)
+			return JsonHandler.fehler("Falsche Anzahl an Parametern.");
+		for (String attribute : attributes) {
+			if (attribute.split("=").length != 2)
+				return JsonHandler.fehler("Parameter sind falsch formatiert.");
+		}
+		String auth = attributes[0].split("=")[1];
+		String pn = attributes[1].split("=")[1];
+		String arbeitsgruppenID = attributes[2].split("=")[1];
+		if (pn.length() != 12)
+			return JsonHandler.fehler("Ungültige Personalnummer.");
+		if (!Utils.checkIfArbeitsgruppeExistsFromID(arbeitsgruppenID))
+			return JsonHandler.fehler("Arbeitsgruppe existiert nicht.");
+		if(!Utils.checkIfMitarbeiterExists(pn))
+			return JsonHandler.fehler("Der Mitarbeiter existiert nicht.");
+		if(Utils.isInArbeitsgruppe(arbeitsgruppenID, pn))
+			return JsonHandler.fehler("Der Mitarbeiter ist bereits in der Arbeitsgruppe vorhanden.");
+		if (!SessionHandler.checkSessionID(auth))
+			return JsonHandler.fehler("SessionID ist ungültig.");
+		// TODO checkRights
+		String sqlStmt = "INSERT INTO gfos.arbeitsgruppenteilnahme (ArbeitsgruppenID, Mitarbeiter) Values ('"+ arbeitsgruppenID +"', '" + pn + "');";
+		try {
+			int rs = QueryHandler.update(sqlStmt);
+			if(rs == 0)
+				return JsonHandler.fehler("Das Einfügen des Mitarbeitsers konnte nicht durchgeführt werden.");
+			return JsonHandler.erfolg("Der Mitarbeiter wurde in die Arbeitsgruppe eingefügt.");
+		} catch (SQLException e) {
+			return JsonHandler.fehler(e.toString());
+		}
+	}
 
 }
